requirements.txt
beautifulsoup4==4.5.1
Django==1.10.4
django-allauth==0.29.0
django-bootstrap3==7.1.0
django-compressor==2.1
django-cors-headers==1.3.1
django-crispy-forms==1.6.1
django-filter==1.0.1
django-modelcluster==2.0
django-taggit==0.18.3
django-treebeard==4.0.1
djangorestframework==3.5.3
html5lib==0.999999999
Pillow==3.4.2
psycopg2==2.6.2
pytz==2016.7
simplejson==3.10.0
scrapy==1.2.2
Unidecode==0.4.19
wagtail==1.7
webencodings==0.5
wheel==0.24.0
Willow==0.4
